send resuest_status data to the backend
